fix: sessions with similar name as current are no longer excluded
Consider the following 3 sessions: foo, foo-bar, baz.
Prior to this fix, while foo is the current session, foo-bar would not
show up in the session list. Similarly, while foo-bar the current
session, foo would not show up in the list. To switch to those missing
sessions, the user would have to scroll up to the zoxide path for the
session or fuzzy search it. This is makes it difficult to quickly jump
back and forth between two or more similarly named sessions.

The cause of this bug is in the logic that controls finding and
excluding the current session from the list of sessions. It uses grep,
but the match is too permissive, resulting in partial matches as well.

This fix updates the grep command to look for an exact match in the
session name, ensuring other sessions with the current session name as a
substring of the name will still correctly display.